Add tests for creating new users
We add several tests to check that new users can be created properly.
Nothing special in that idea, but as part of the process, we added a
helper function to the testing framework that checks for alert
messages that are of the proper format named have_alert().  That
method takes two parameters that include the alert type and any
options that need to be passed on to have_selector().  The
implementation is yet to be completed since the student does not
currently know about flash messages and therefore what the content of
the have_selector() call should even look like.  That will be
completed in future commits.