Better UI for failure on User creation
We now display our error messages and put them next to the form that
generated them.  This makes it much easier for a user to be able to
fix the problem since they see what the error is and have that
information on the same web page as the form they need to fill out.