Test for viewing non-existant user
We create a new test that tries to access a user that doesn't exist.
This introduces the :request type of test and the redirect_to() method
on the response object.  We give a hard-coded user ID of -1 that we
expect will not match any users in the database, so this should always
encounter a failure mode.