New tests for editing a user.
These are a large set of tests that need to be passed to allow a user
to be edited.  They are still not perfect because they assume that
anyone can edit the users, but we don't have a mechanism for logging
in and out, yet, so this will have to do.