Only allow users to edit their own profiles
This commit passes the tests from the previous commit that ensured
that users were only allowed to edit their own profiles.  We had to
change some tests to be able to continue passing with this new
functionality added since they didn't log in to a user before editing
the profile.

Also, a common login file was placed in spec/support/login.rb to allow
tests to quickly allow a user to login without repeating code to do
the same functionality.

Finally, a bug in the current_user method was fixed because it was
passing a nil to User.find if no one was logged in.  This code only
passed a non-nil value to User.find, otherwise it returned a nil to
the calling context.

Conflicts:
	app/controllers/users_controller.rb
	app/views/users/edit.html.erb