Final tests for User redirection
This is the final set of tests that check that the user is properly
redirected when they try to perform an action that they shouldn't be
authorized to do.  The two examples in this commit are trying to
create a user while being logged in and the admin user deleting
themself.