Only allow church managers to edit church profiles
This tests that only church managers are allowed to be able to access
the profiles of the various church profile pages that relate to
edits/updates.  We don't support creating them, yet.  So this is an
initial security step once they are created to ensure they're not
modified by others.