Complete Church test (1st draft)
This should be close to a complete set of tests for the Church
resource.  It tests that we can do any of the CRUD operations on the
resource and that it is robust to various permissions.  The only
things we probably haven't checked is who can delete a church and
whether services are deleted when an associated church is.  All the
same, this is close to complete.

Conflicts:
	spec/factories.rb