Test that Church displays all attendees.
This checks that a church lists all the people that are marked as
attendees of the church on its profile page.

BIG NOTE:  This commit includes some fundamental changes to the church
and user models so that Rails will be able to understand the
relationships between the two models since there are two relationships
there: church/attendee and church/manager.  We need to disambiguate
them because they both use church (singular) to connect from the user
for both relationships.