Allow a user to claim a ride
This tests that when a user views a ride that is provided that they
can see the information necessary to decide if they want to take the
ride and then perform the action of claiming the ride for themselves.

Much error testing is missing as I am simply focusing on core
functionality first before ensuring all the other stuff works in every
case.