Move date to future (barely)
This test was failing because the hard-coded date was no longer in the
future.  This commit moves it one year in the future and will have the
same failure if not fixed more robustly.  Said "good" fix is not being
committed now because I don't have Rails updated on my current machine
and do not have time to get that working while students wait for it to
be changed.  So...quick and dirty hack so that I can remember the change
and don't let it linger too long.