Randomize hard-coded sizes
Turns out that when you hard code a value in a test, that people will
hard code the value in their HTML output to match it instead of figuring
out what the value means for the program and outputs that meaningful
value instead.  This change moves all those hard-coded sizes to be
randomized over a small range so that previous "correct solutions" are
now marked as broken like they always should have been.