Change redirect on editing non-existent Ride
Originally, the test tried to redirect to rides_path, but that isn't a
legal path since rides resources are nested in services.  Unfortunately,
we don't have a service to redirect back to because the edit action
doesn't take a service ID and we don't have a legal Ride to look up the
nesting service.  As a result, we route to the only place that makes
sense:  root_path.  After all, which service?  which church?  Lame, but
no better option presents itself.