Add additional authorization checks for Rides & Churches
There were missing authorization checks for Rides to reflect the new
functionality added with the previous tests.  While we were in there, we
made some quite minor changes to the Church checks, too.  Nothing too
drastic, but keeps it symmetric with User authorization checks.
Basically, these new checks ensure that only a logged in user can try to
directly edit the ride data.  This check does not check that any user
can claim a ride because I have not codified which controller is
responsible for handling the claim.  Some students have put that in the
user controller and others have put it in the ride controller.
Furthermore, I have not stated how that data is passed from the view to
the respective controller, so it would be difficult for current students
to be able to modify their code to reflect those changes in these tests.
As a result, we only check that a "normal" update can only be done with
the proper user.  We don't check that a claim is passed through because
we don't know how to make that claim for every student's implementation.